+++ controllers/lab/membersController.js
@@ -1,12 +1,38 @@
 const db = require('../../models')
 const Members = db.Members
 const errorHandler = require('../../middleware/errorHandler')
+const { Op } = require('@sequelize/core')
 
 
 class MembersController{
     addMember = async(req, res) => {
         try {
             const { tag, name, researchDirection, mail, paperTopic } = req.body
+
+            if (!tag || !name || !researchDirection || !mail){
+                return res.status('400').send(errorHandler.contentEmpty())
+            }
+            if (tag == "alumnus"){
+                if (!paperTopic){
+                    return res.status('400').send(errorHandler.contentEmpty())
+                }
+            }
+
+            const checkExist = await Members.findOne({
+                where: {
+                    [Op.or]: [
+                        {name: name},
+                        {mail :mail}
+                    ]
+                }
+            })
+
+            if (checkExist){
+                return res.status('409').send(errorHandler.userAlreadyExist())
+            }
+
+
+
             let infor = {
                 tag: tag,
                 name: name,
@@ -16,9 +42,10 @@ class MembersController{
                 paperTopic: paperTopic
 
             }
-            const upload = await Members.create(infor)
+            const data = await Members.create(infor)
             return res.status('200').send({
-                message: upload
+                message: `Create ${data.name} sucessfully!`,
+                detail: data 
             })
            }
         catch (error) {
@@ -41,6 +68,39 @@ class MembersController{
     updateMember = async(req, res) => {
         try {
             const { tag, name, researchDirection, mail, paperTopic } = req.body
+
+            if (!tag || !name || !researchDirection || !mail){
+                return res.status('400').send(errorHandler.contentEmpty())
+            }
+            if (tag == "alumnus"){
+                if (!paperTopic){
+                    return res.status('400').send(errorHandler.contentEmpty())
+                }
+            }
+
+            const checkExist = await Members.findOne({
+                where:{
+                    id: req.params.id
+                }
+            })
+
+            if (!checkExist){
+                return res.status('404').send(errorHandler.dataNotFind())
+            }
+
+            const checkInforExist = await Members.findOne({
+                where: {
+                    [Op.or]: [
+                        {name: name},
+                        {mail :mail}
+                    ]
+                }
+            })
+
+            if (checkInforExist){
+                return res.status('409').send(errorHandler.userAlreadyExist())
+            }
+
             let infor = {
                 tag: tag,
                 name: name,
@@ -56,7 +116,7 @@ class MembersController{
                 }
             })
             return res.status('200').send({
-                message: upload
+                message: "Updated Sucessfully!"
             })
            }
         catch (error) {
@@ -67,13 +127,17 @@ class MembersController{
         }
 
     deleteMember = async(req, res) =>{
-        const id = req.params.id
         try {
-            const data = await Members.findOne({
+            const checkExist = await Members.findOne({
                 where:{
                     id: req.params.id
                 }
             })
+
+            if (!checkExist){
+                return res.status('404').send(errorHandler.dataNotFind())
+            }
+
             const del = await Members.destroy({
                 where: {id : req.params.id}
             })
